HBASE-22324 loss a mass of data when the sequenceId of cells greater than Integer.Max

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/MemStoreCompactorSegmentsIterator.java
@@ -56,7 +56,7 @@ public class MemStoreCompactorSegmentsIterator extends MemStoreSegmentsIterator 
     super(compactionKVMax);
 
     List<KeyValueScanner> scanners = new ArrayList<KeyValueScanner>();
-    AbstractMemStore.addToScanners(segments, Integer.MAX_VALUE, scanners);
+    AbstractMemStore.addToScanners(segments, Long.MAX_VALUE, scanners);
     // build the scanner based on Query Matcher
     // reinitialize the compacting scanner for each instance of iterator
     compactingScanner = createScanner(store, scanners);
